+++ pymilo/chains/neural_network_chain.py
@@ -70,7 +70,7 @@ def transport_neural_network(request, command, is_inner_model=False):
                     'object': request,
                 })
 
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         try:
             return deserialize_neural_network(request, is_inner_model)
         except Exception as e:
@@ -118,7 +118,7 @@ def deserialize_neural_network(neural_network, is_inner_model=False):
 
     for transporter in NEURAL_NETWORK_CHAIN:
         NEURAL_NETWORK_CHAIN[transporter].transport(
-            neural_network, Command.DESERIALZIE, is_inner_model)
+            neural_network, Command.DESERIALIZE, is_inner_model)
     for item in data:
         setattr(raw_model, item, data[item])
     return raw_model
@@ -144,7 +144,7 @@ def _validate_input(model, command):
                     'object': model
                 }
             )
-    elif command == Command.DESERIALZIE:
+    elif command == Command.DESERIALIZE:
         if is_neural_network(model.type):
             return
         else: